게시판 수정, 삭제 완료 댓글 수정, 삭제 완료 그룹 생성 완료
------해야할 일------

그룹 수정, 삭제

+++ ggirick_client_backend/src/main/java/com/kedu/ggirick_client_backend/services/board/BoardGroupService.java
@@ -26,10 +26,10 @@ public class BoardGroupService {
         return boardGroupDAO.getGroupOwner(groupId);
     }
 
-    public void updateGroupEmployee(List<String> members, int boardId) {
+    public void updateGroupEmployee(List<String> members, int boardId, String ownerId) {
         // 없던 멤버 추가
         for (String member : members) {
-            if(!getGroupEmployeeList(boardId).contains(member)) {
+            if(!getGroupEmployeeList(boardId).contains(member) && !member.equals(ownerId)) {
                 Map<String, Object> params = new HashMap<>();
                 params.put("boardId", boardId);
                 params.put("userId", member);
@@ -39,12 +39,31 @@ public class BoardGroupService {
 
         // 있던 멤버 삭제
         for (String existedMember : getGroupEmployeeList(boardId)) {
-            if(!members.contains(existedMember)) {
+            if(!members.contains(existedMember) && !existedMember.equals(ownerId)) {
                 Map<String, Object> params = new HashMap<>();
                 params.put("boardId", boardId);
                 params.put("userId", existedMember);
                 boardGroupDAO.deleteGroupEmployee(params);
             }
         }
+    }
+
+    public void addBoardGroup(BoardGroupDTO groupInfo) {
+        int groupId = boardGroupDAO.addBoardGroup(groupInfo);
+
+        Map<String, Object> params = new HashMap<>();
+
+        params.put("boardId", groupId);
+        params.put("userId", groupInfo.getOwnerId());
+
+        boardGroupDAO.insertGroupEmployee(params);
+    }
+
+    public void deleteBoardGroup(int groupId) {
+        boardGroupDAO.deleteBoardGroup(groupId);
+    }
+
+    public void updateBoardGroup(BoardGroupDTO groupInfo) {
+        boardGroupDAO.updateBoardGroup(groupInfo);
     }
 }
